Fix incorrect import due to refactoring of observers
